+++ src/services/adex-validator/actions.js
@@ -1,6 +1,7 @@
 import Requester, { handleRequesterErrorRes } from 'services/requester'
 import { getEthers } from 'services/smart-contracts/ethers'
 import { getSigner } from 'services/smart-contracts/actions/ethers'
+import dateUtils from 'helpers/dateUtils'
 import ewt from './ewt'
 
 const BEARER_PREFIX = 'Bearer '
@@ -143,6 +144,7 @@ export const identityAnalytics = async ({
 	segmentByChannel,
 	start,
 	end,
+	weekHoursSpan,
 }) => {
 	const baseUrl = ANALYTICS_DATA_VALIDATOR_URL
 	const requester = getValidatorRequester({ baseUrl })
@@ -159,6 +161,8 @@ export const identityAnalytics = async ({
 				eventType,
 				start,
 				end,
+				weekHoursSpan,
+				timezone: dateUtils.getCurrentTimezone(),
 			},
 			headers: {
 				authorization: BEARER_PREFIX + leaderAuth,